chore: updated with dropdown with currency list
