+++ src/App.tsx
@@ -1,14 +1,19 @@
 import { useForm } from "react-hook-form";
 import { ExchangeRates } from "./types";
 import { formatMoney } from "./utils";
-import useFetchLatestRates from "./hooks/useFetchLatestRates";
+import { useFetchLatestRates, useFetchCurrencyList } from "./hooks";
 
 const App = () => {
-  const { exchangeRates, setAmount } = useFetchLatestRates();
+  const { exchangeRates, setAmount, setFromCurrency } = useFetchLatestRates();
+  const { currencyList } = useFetchCurrencyList();
   const { register, handleSubmit } = useForm();
 
-  const onSubmit = (data) => {
+  const onSubmit = (data: {
+    amount: number;
+    fromCurrency: keyof typeof currencyList;
+  }) => {
     setAmount(data.amount);
+    setFromCurrency(data.fromCurrency);
   };
 
   return (
@@ -17,7 +22,19 @@ const App = () => {
       <form onSubmit={handleSubmit(onSubmit)}>
         <label htmlFor="amount">Amount</label>
         <input id="amount" type="number" {...register("amount")} />
-        <button>Convert</button>
+        <label htmlFor="fromCurrency">Currency</label>
+        <select {...register("fromCurrency")} defaultValue="EUR">
+          {currencyList?.map(({ currency }) => (
+            <option
+              key={currency}
+              value={currency}
+              selected={currency === "EUR"}
+            >
+              {currency}
+            </option>
+          ))}
+        </select>
+        <input type="submit" />
       </form>
       <div>
         {exchangeRates?.length > 0 ? (
