+++ src/types.ts
@@ -1,43 +1,17 @@
-export const Currencies = {
-  AUD: "Australian Dollar",
-  BGN: "Bulgarian Lev",
-  BRL: "Brazilian Real",
-  CAD: "Canadian Dollar",
-  CHF: "Swiss Franc",
-  CNY: "Chinese Renminbi Yuan",
-  CZK: "Czech Koruna",
-  DKK: "Danish Krone",
-  EUR: "Euro",
-  GBP: "British Pound",
-  HKD: "Hong Kong Dollar",
-  HUF: "Hungarian Forint",
-  IDR: "Indonesian Rupiah",
-  ILS: "Israeli New Sheqel",
-  INR: "Indian Rupee",
-  ISK: "Icelandic Króna",
-  JPY: "Japanese Yen",
-  KRW: "South Korean Won",
-  MXN: "Mexican Peso",
-  MYR: "Malaysian Ringgit",
-  NOK: "Norwegian Krone",
-  NZD: "New Zealand Dollar",
-  PHP: "Philippine Peso",
-  PLN: "Polish Złoty",
-  RON: "Romanian Leu",
-  SEK: "Swedish Krona",
-  SGD: "Singapore Dollar",
-  THB: "Thai Baht",
-  TRY: "Turkish Lira",
-  USD: "United States Dollar",
-  ZAR: "South African Rand",
-} as const;
+import { Currencies } from "./constants";
+
+export type ExchangeRates = {
+  currency: keyof typeof Currencies;
+  rate: number;
+};
 
 export type CurrencyList = {
   currency: keyof typeof Currencies;
   name: string;
 };
 
-export type ExchangeRates = {
-  currency: keyof typeof Currencies;
-  rate: number;
+export type FormatUrl = {
+  amount: number;
+  fromCurrency: keyof typeof Currencies;
+  toCurrency: keyof typeof Currencies;
 };
